+++ package.json
@@ -44,12 +44,12 @@
   "dependencies": {
     "@apollo/react-hooks": "^3.1.5",
     "@babel/polyfill": "^7.0.0",
-    "@jahia/data-helper": "^0.0.26",
+    "@jahia/data-helper": "^1.0.0",
     "@jahia/design-system-kit": "^1.1.2",
     "@jahia/icons": "^1.1.1",
     "@jahia/moonstone": "^0.16.0",
-    "@jahia/react-material": "^2.1.1",
-    "@jahia/ui-extender": "^0.5.4",
+    "@jahia/react-material": "^3.0.1",
+    "@jahia/ui-extender": "^1.0.0",
     "@material-ui/core": "^3.6.2",
     "@material-ui/icons": "^3.0.2",
     "@material-ui/lab": "^3.0.0-alpha.16 ",
